+++ 02-popular-movies/app/src/main/java/com/esanz/nano/movies/ui/adapter/SimpleViewHolder.java
@@ -1,4 +1,4 @@
-package com.esanz.nano.movies.ui.widget;
+package com.esanz.nano.movies.ui.adapter;
 
 import android.support.annotation.IdRes;
 import android.support.v7.widget.RecyclerView;
