+++ 02-popular-movies/app/src/main/java/com/esanz/nano/movies/ui/viewModel/MovieDetailViewModel.java
@@ -1,0 +1,6 @@
+package com.esanz.nano.movies.ui.viewModel;
+
+import android.arch.lifecycle.ViewModel;
+
+public class MovieDetailViewModel extends ViewModel {
+}
