+++ 02-popular-movies/app/src/main/java/com/esanz/nano/movies/ui/viewModel/MovieListViewModel.java
@@ -1,14 +1,12 @@
-package com.esanz.nano.movies.ui;
+package com.esanz.nano.movies.ui.viewModel;
 
 import android.arch.lifecycle.MutableLiveData;
 import android.arch.lifecycle.ViewModel;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import com.esanz.nano.movies.MovieApplication;
 import com.esanz.nano.movies.repository.MovieDataSource;
 import com.esanz.nano.movies.repository.MovieRepository;
-import com.esanz.nano.movies.repository.dao.FavoriteDao;
 import com.esanz.nano.movies.repository.model.Movie;
 import com.esanz.nano.movies.repository.model.PaginatedMovieResponse;
 import com.esanz.nano.movies.utils.MovieConstant;
@@ -16,7 +14,7 @@ import com.esanz.nano.movies.utils.MovieConstant;
 import java.util.List;
 import java.util.Objects;
 
-public class MovieViewModel extends ViewModel {
+public class MovieListViewModel extends ViewModel {
 
     public final MutableLiveData<List<Movie>> movieListLiveData = new MutableLiveData<>();
 
@@ -27,7 +25,7 @@ public class MovieViewModel extends ViewModel {
 
     private int sortType = MovieConstant.SortType.RATING;
 
-    public MovieViewModel(@NonNull final MovieRepository movieRepository) {
+    public MovieListViewModel(@NonNull final MovieRepository movieRepository) {
         this.movieRepository = Objects.requireNonNull(movieRepository);
     }
